added graphs (currently gives random numbers)

+++ src/main/java/engine/math/Vector3.java
@@ -2,7 +2,7 @@ package engine.math;
 
 public class Vector3 {
 
-    public float x, y, z; // this is 2d, however z is used if it is above or below
+    public double x, y, z; // this is 2d, however z is used if it is above or below
 
     public Vector3() {
         x = 0;
@@ -10,7 +10,7 @@ public class Vector3 {
         z = 0;
     }
 
-    public Vector3(float x, float y, float z) {
+    public Vector3(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
